test: speed up live sandbox e2e suite
